Add attempt to redisplay with hook.
However, it doesn't work. It's too much overload and too slow. I think
this is the hardest part of this project. When the state changes, I need
to be able to calculate the smallest diff and only re-display that part.

It seems that we must start within the state: We cannot re-read the
whole state externally all the time.